+++ changelog.md
@@ -1,5 +1,11 @@
 # contentful-pull changelog
 
+#### v0.2
+  - `.get`, `.sync` now has a "resolveLinks"-option which resolves entry and asset links
+  - Provide `{raw: true}` in sync or get methods to get the raw data from contentful
+  - Now saving data as it comes from Contentful and processing it upon `.get`
+  - Sync now works with updates
+
 #### v0.1
   - `.get`, `.sync`
   - Saves to file and reformats data to a workable format